ML-backend/voice-agent-backend/requirements.txt
flask==3.0.2
deepgram-sdk==4.0.0
python-dotenv==1.0.1
PyAudio==0.2.14
requests==2.32.3
flask-cors==4.0.0
flask-socketio==5.3.6
pytest==8.0.0
anthropic==0.35.0
httpx==0.27.2